# This is a combination of 7 commits. parent a701ee3c5a6395fc87a4b928672346fff398a8ba author red-hat-trusted-app-pipeline <123456+red-hat-trusted-app-pipeline[bot]@users.noreply.github.com> 1698935322 +0000 committer Kyl-Bempah <24903083+Kyl-Bempah@users.noreply.github.com> 1699984168 -0500
# This is a combination of 2 commits.
# This is the 1st commit message:

chore(deps): update rhtap references

# This is the commit message #2:

Enable multi-arch builds

Signed-off-by: Kyl-Bempah <24903083+Kyl-Bempah@users.noreply.github.com>
# This is the commit message #3:

OCPBUGS-16079: No error for overlapping service network and API IP

Fixes: OCPBUGS-16079

Ensure that the nodeport IP for the API Server does not conflict
with the Service Network CIDR.

# This is the commit message #4:

Stop exposing kas on 6443 private route service load balancer

We want to remove the unncessary exposure of port 6443 externally, as this increases the used SG rules, which decreases our max number of HC per management aws account #3139 We split the PR to roll out the HO first to all envs to make sure the haproxy in dataplane for newly created HCs don't point to a port 6443 that won't be exposed.

# This is the commit message #5:

OCPBUGS-10423: Update regex validation for nodepool.spec.taints.value

# This is the commit message #6:

HOSTEDCP-1278: Adjustment cluster-cidr,service-cidr to support dual-stack

# This is the commit message #7:

KubeVirt: create the etcd encryption key secret, if missing

To allow creating of KubeVirt hosted cluster using the hosted cluster
API (rather than using the cli).

When creating the hosted cluster using the cli, the cli also creates the
secret. But when creating the hosted cluster using the hosted cluster
API, the secret is not created.

This PR changes hypershift so it now creates the etcd
encryption key secret, if it is not already exist.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>